// server/sonar-web/src/main/js/types/types.ts
export interface Rule {
  cleanCodeAttribute?: CleanCodeAttribute;
  cleanCodeAttributeCategory?: CleanCodeAttributeCategory;
  impacts: SoftwareImpact[];
  isTemplate?: boolean;
  key: string;
  lang?: string;
  langName?: string;
  name: string;
  params?: RuleParameter[];
  severity: string;
  status: string;
  sysTags?: string[];
  tags?: string[];
  type: RuleType;
}

export interface RestRule {
  cleanCodeAttribute?: CleanCodeAttribute;
  cleanCodeAttributeCategory?: CleanCodeAttributeCategory;
  impacts: SoftwareImpact[];
  key: string;
  language?: string;
  languageName?: string;
  name: string;
  parameters?: RestRuleParameter[];
  severity: string;
  status: string;
  systemTags?: string[];
  tags?: string[];
  template?: boolean;
  type: RuleType;
}

export interface RuleActivation {
  createdAt: string;
  impacts: { severity: SoftwareImpactSeverity; softwareQuality: SoftwareQuality }[];
  inherit: RuleInheritance;
  params: { key: string; value: string }[];
  prioritizedRule: boolean;
  qProfile: string;
  severity: string;
}
// ...
export type RuleInheritance = 'NONE' | 'INHERITED' | 'OVERRIDES';

export interface RestRuleParameter {
  defaultValue?: string;
  htmlDescription?: string;
  key: string;
  type: string;
}

export interface RuleParameter {
  defaultValue?: string;
  htmlDesc?: string;
  key: string;
  type: string;
}

export type RuleScope = 'MAIN' | 'TEST' | 'ALL';

export const RuleTypes = [
  'BUG',
  'VULNERABILITY',
  'CODE_SMELL',
  'SECURITY_HOTSPOT',
  'UNKNOWN',
] as const;
export type RuleType = (typeof RuleTypes)[number];